requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
tensorflow==2.16.1
numpy==1.24.3
Pillow==10.0.0
watchdog==3.0.0
gunicorn==21.2.0
